Split subtract.test.js into subtractGeom2.test.js and subtractGeom3.test.js
